[fix]: use weak reference to internal sinks when vending to clients (#189)
previously we would allow a strong reference to the internal `ReusableSink` to be captured by the `Sink` types that are vended to clients. this meant that events sent into such 'external' sinks would still propagate through the event handling machinery a bit before realizing that the backing workflow node to which they were destined was no longer around. by using weak references we can prevent extending the life of the internal sinks (and underlying EventPipe) unnecessarily in such cases.

- make `ReusableSink` captures `weak` when vending to the 'outside world'
- update various access control values to facilitate testing this behavior