[feat]: add experimental WorkflowUI telemetry hooks (#221)
### Motivation

we'd like to be able to build out some observation tooling that is aware of certain events that occur within WorkflowUI. specifically, view lifecycle data is of particular interest. currently the options for tying into UIViewController lifecycle events are either:

1. introduce new subclasses in client code that override the methods of interest and provide observation hooks there.
2. swizzle the methods of interest to allow client observation

option `1.` may place a large burden on consumers that have many existing subclasses (of, say, ScreenViewController), and only works for the `open` types defined in `WorkflowUI`, which excludes `WorkflowHostingController` & `DescribedViewController`. option `2.` may require less code to implement, but adds indirection (swizzling is often not particularly discoverable), runtime overhead, and may not work for Swift-only methods.

note: this is all currently implemented behind 'experimental' SPI declarations. the intent would be to treat this as sort of an 'alpha' version of this functionality, and to set expectations that it may have breaking changes going forward, without a major version increment in the library.

### Implementation

to achieve this goal, we've implemented an approach that does the following:

- unifies all `UIViewController` subtypes defined in `WorkflowUI` under a new parent class `WorkflowUIViewController`
- instruments some view controller lifecycle events (viewDidAppear, etc) within the shared base class
- exposes a static value that can be used to register a 'global' observer of these events
- adds some new protocols and event data types to enable observation of WorkflowUI lifecycle events