[feat]: add 'screen container' API (#258)
## Motivation

in some circumstances it is helpful to know when a type is serving as a logical 'container' for a rendering, and in particular a container of a `Screen`. the types currently exposed by WorkflowUI which serve this purpose are the following:

1. `AnyScreen`
1. `WorkflowHostingController<ScreenType, Output>`
1. `ScreenViewController<ScreenType>`

since both the hosting controller & screen view controller are generic, there is no way to dynamically check if an arbitrary type is an instance without also specifying the generic parameters. this makes certain types of dynamic introspection clunky or impossible depending on code structure. to illustrate the point, consider the following somewhat contrived example:

```swift
func makeUninspectableScreenVC() -> UIViewController {
    struct LocalScreen: Screen { ... } // `LocalScreen` symbol is only visible in this function
    return ScreenViewController(screen: LocalScreen(), environment: .empty)
}

let isSVC = makeUninspectableScreenVC() is ScreenViewController<???> // no generic can be specified here that makes this true
```
we'd like our logical 'Screen container' types to share an interface that allows access to the underlying screen information, which can help with debugging and other runtime data collection.

## Changes

- adds the `ScreenContaining` protocol. this has a single requirement that conformances must provide a `containedScreen` value, returning an `any Screen` type that is the semantically 'contained' screen
- adds conformances to this protocol to the existing screen containers in WorkflowUI
- adds a utility protocol extension to traverse a chain of containers to find the 'leaf' screen