feat: flexible SwiftUI preferredContentSize calc [UI-7665] (#320)
This is a possible solution for preferredContentSize of SwiftUI content,
for discussion.

A summary of the problem:
- The system provided PCS calculation is unconstrained and not useful to
us. We need to calculate this manually.
- SwiftUI ScrollViews greedily fill the space available, but will report
the natural content size when given an infinite size.

This solution finds an appropriate PCS for the general case by measuring
in a few different constraints and combining the results. This
accommodates content with no scroll view, as well as content with scroll
views in either axis or both.