Rename `expect(worker:producingOutput:)` to be more clear about what it is asserting (#323)
When I first saw tests that used this function, I thought it was
asserting that the `worker` passed in, matches the `workflow.worker`
(which it does do), and that it was asserting that `producingOutput ==
workflow.worker.output` (which it does NOT do). Instead, neither of the
workers are ever run. The `producingOutput` is used as mock output. So
renaming these to make it more clear what is going on.


## Checklist

- [x] Unit Tests
- [x] UI Tests
- [x] Snapshot Tests (iOS only)
- [x] I have made corresponding changes to the documentation