Adds a convenience for mapping Never outputs to other types (#340)
Adds a way to map a workflow's `Never` output type to a different type
without generating compiler warnings. Fixes
https://github.com/square/workflow-swift/issues/339

I like this solution because the natural inclination is to simply delete
the block containing unreachable code and that just works:

```diff
-            .mapOutput { switch $0 {} }
+            .mapOutput()
```

I'm happy to add a test using this but the true validation is to
successfully build without warnings.