[chore]: update CI config (#343)
updates the CI config with the following changes:

1. bump to Xcode 16.1
2. add a test matrix for the unit tests
i. this is largely copied from the analogous logic in the Blueprint CI
[config](https://github.com/square/Blueprint/blob/067015242d815f698550b3960353f545beea7183/.github/workflows/tests.yaml)
    ii. runs tests against the 3 most recent iOS SDKs (16.4, 17.5, 18.1)
3. split out snapshot tests to run only against iOS 18.1
i. this was primarily due to not knowing how to configure them to
function correctly with the test matrix – something that should
presumably be improved, but is no worse than what we're currently doing.
4. disables a failing unit test on iOS 18+
i. TODO: re-enable by adding support for app-hosted unit tests (tracked
via [this issue](https://github.com/square/workflow-swift/issues/346))