[feat]: add runtime option to skip render if state is inferred to be unchanged (#364)
adds support to conditionally opt in a WorkflowHost runtime to use a
render-skipping optimization that checks if we can reasonably infer that
no workflow node state was changed during a series of action
applications. if the inference indicates no state has changed, the
render pass will be skipped. the changes made to support this include:

- adding a new `ActionApplicationResult` type to model the result of
applying an action to a node. this type carries some invalidation state
information with it to propagate up the tree.
- adding new conditional logic into the `applyAction()` method (renamed
from `openAndApply()`) to conditionally try to infer if state has
changed under the following conditions:
- if the `State` instance can be dynamically inferred to be `Equatable`,
then the before & after values will be compared to determine if the
update should be treated as invalidating state or not
- if the type of `State` is equivalent to `Void`, then it will be
treated as never having changed
- if the sequence of action applications makes it to the root node and
the final `ActionApplicationResult` indicates that no state was changed,
the root node will skip the subsequent render pass. note: if an output
event is produced, that will still be delivered and is independent of
this logic.

this change has the potential to alter observable behavior, and so is
opt-in behind a conditional runtime configuration option.

primary changes:
- `WorkflowNode.applyAction()`
- `WorkflowHost.handle()`