ci: update macOS and Xcode (#373)
Bumps Xcode to 16.4 on the Package.swift tests and pins images to
macos-15 so they're not moving underneath us.

To do this bump I had to remove the (redundant) unit test definitions in
Package.swift. The reason is that `swift test` won't build some of our
tests that depend on Perception due to availability (`error: 'Bindable'
is unavailable in macOS: Use @Bindable without the 'Perception.'
prefix`) and I cannot find a way to selectively execute tests without
building all of them. Tuist still does not support creating a test
target for an external macro.

All tests are still run in CI: unit tests on iOS via Tuist, macros on
macOS via `swift test`.