chore: Bump Perception to 2.0.6 (#377)
## Checklist

**Bump perception to 2.0.6**
**`tuist install --path Samples` to update Sample resolved
dependencies**

The Perception version specified in Package.swift is now
`1.5.0"..<"3.0.0`. This allows easier upgrade and downgrades by
consumers.

I've opened this in favor of
https://github.com/square/workflow-swift/pull/376 because Perception has
been [updated](https://github.com/pointfreeco/swift-perception/pull/156)
following Apple's reversion of
https://github.com/swiftlang/swift/pull/83436 and it avoids test updates
on our end for now.

I didn't run the `tuist install` with the `--update` flag this time
either. There doesn't seem to be a way to scope the updates to only
Perception and it's sub-dependencies and I didn't want all the other
clutter associated with this change.

- [x] Unit Tests
- [x] UI Tests
- [x] Snapshot Tests (iOS only)
- [x] I have made corresponding changes to the documentation
- [x] Resolve new Perception warning in ios-register:
https://github.com/squareup/ios-register/pull/131811
- [x] Resolve new instances of Perception warnings in Market:
https://github.com/squareup/market/pull/11042

---------

Co-authored-by: Andrew Watt <100192+watt@users.noreply.github.com>