[refactor]: allow default runtime config to be set more than once (#385)
this changes the default runtime configuration logic so that it no
longer tries to enforce that it can only be updated once during the
execution of a process. in practice we may want to update this for all
hosts created going forward from some point in time, based on a feature
flag or something.